feat(invitation_code): add endpoints for generating and retrieving available invitation codes
