+++ app/api/routes/auth.py
@@ -1,8 +1,7 @@
-from fastapi import APIRouter, Header
+from fastapi import APIRouter
 
-from app.core.config import settings
 from app.exceptions.auth_exceptions import AuthException
-from app.models.auth import CheckInvitationCodeRequest, CheckInvitationCodeResponse, GenerateInvitationCodeResponse
+from app.models.auth import CheckInvitationCodeRequest, CheckInvitationCodeResponse
 from app.utils.invitation_code_utils import InvitationCodeUtils
 from app.utils.jwt_utils import JWTUtils
 
@@ -39,19 +38,4 @@ async def check_invitation_code(data: CheckInvitationCodeRequest) -> CheckInvita
     return CheckInvitationCodeResponse(access_token=access_token)
 
 
-@router.post("/generate-invitation-code", response_model=GenerateInvitationCodeResponse)
-async def generate_invitation_code(api_key: str = Header(...)) -> GenerateInvitationCodeResponse:
-    """
-    Check if the invitation code is valid and return an access token
 
-    Args:
-        api_key (str): The API key to authenticate
-
-    Returns:
-        GenerateInvitationCodeResponse: Contains InvitationCode
-    """
-    # Check if the API key is valid
-    if api_key != settings.ADMIN_API_KEY:
-        raise AuthException(status_code=401, message="Invalid APIKEY")
-    invitation_code = invitation_code_utils.generate_invitation_code()  # Generate a new invitation code
-    return GenerateInvitationCodeResponse(invitation_code=invitation_code.code)
